+++ lib/reactive_forms.dart
@@ -8,6 +8,7 @@ export 'src/exceptions/form_array_invalid_index_exception.dart';
 export 'src/exceptions/form_builder_invalid_initialization_exception.dart';
 export 'src/exceptions/form_control_not_found_exception.dart';
 export 'src/exceptions/form_control_parent_not_found_exception.dart';
+export 'src/input_parsers/input_parsers.dart';
 export 'src/models/abstract_control.dart';
 export 'src/models/control_status.dart';
 export 'src/models/form_array.dart';